refactor timeline into separate widget view
