+++ src/lib.rs
@@ -2,3 +2,4 @@ pub use makepad_widgets;
 pub mod app;
 pub mod matrix;
 pub mod sliding_sync;
+pub mod timeline;
