+++ src/timeline/mod.rs
@@ -1,0 +1,1 @@
+pub mod timeline_view;
